requirements.txt
aniso8601==2.0.1
apturl==0.5.2
astroid==1.6.1
beautifulsoup4==4.4.1
Brlapi==0.6.4
chardet==2.3.0
chrome-gnome-shell==0.0.0
click==6.7
command-not-found==0.3
configparser==3.5.0
cycler==0.9.0
defer==1.0.6
Flask==0.12.2
Flask-Cors==3.0.3
Flask-RESTful==0.3.6
GDAL==1.11.3
gunicorn==19.7.1
html5lib==0.999
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
language-selector==0.1
lazy-object-proxy==1.3.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==1.0
matplotlib==1.5.1
mccabe==0.6.1
menulibre==2.1.3
numpy==1.11.0
Pillow==3.1.2
psutil==3.4.2
psycopg2-binary==2.7.4
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
pylint==1.8.2
pyparsing==2.0.3
pysmbc==1.0.15.5
python-apt===1.1.0.b1-ubuntu0.16.04.1-zorin1
python-dateutil==2.4.2
python-debian==0.1.27
python-systemd==231
pytz==2014.10
pyxdg==0.25
requests==2.9.1
screen-resolution-extra==0.0.0
sessioninstaller==0.0.0
six==1.11.0
ssh-import-id==5.5
system-service==0.3
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.13.1
usb-creator==0.3.0
virtualenv==15.0.1
Werkzeug==0.14.1
wrapt==1.10.11
xkit==0.0.0
